test(unit): converted tests for the dialect scaffolder to vitest
for #452
